Refactor Projects.jsx into presentational and container components in separate directories
