+++ static/js/components/Delete/index.jsx
