+++ static/js/containers/NewProjectForm/index.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { reduxForm } from 'redux-form';
+
+import ProjectForm from '../../components/ProjectForm';
+import * as Validate from '../../validate';
+
+
+const validate = Validate.createValidator({
+  projectName: [Validate.required],
+});
+
+export default reduxForm({
+  form: 'newProject',
+  fields: ['projectName', 'projectDescription'],
+  validate
+})(ProjectForm);
